requirements.txt
# 知识库自动化管理系统 - Python依赖

# YAML解析（用于配置文件和frontmatter）
pyyaml>=6.0

# 中文分词（用于关键词提取和相似度计算）
jieba>=0.42.1

